Refs #88. Rename xml folder name to xml_r.
