+++ n_body_simulations/plotting/simulation_animator.py
@@ -1,6 +1,6 @@
 # Project Repository : https://github.com/robertapplin/N-Body-Simulations
 # Authored by Robert Applin, 2020
-from n_body_simulations.xml.xml_reader import get_user_interface_property
+from n_body_simulations.xml_r.xml_reader import get_user_interface_property
 
 from matplotlib.animation import FuncAnimation
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
