+++ n_body_simulations/qt/body_data_table.py
@@ -1,7 +1,7 @@
 # Project Repository : https://github.com/robertapplin/N-Body-Simulations
 # Authored by Robert Applin, 2020
 from n_body_simulations.qt.table_item_delegates import ColourItemDelegate, DoubleItemDelegate, StringItemDelegate
-from n_body_simulations.xml.xml_reader import get_user_interface_property
+from n_body_simulations.xml_r.xml_reader import get_user_interface_property
 
 from PyQt5.QtCore import pyqtSignal, QEvent, QModelIndex, QObject, QPersistentModelIndex, Qt
 from PyQt5.QtWidgets import QAbstractItemView, QTableWidget, QTableWidgetItem
