+++ n_body_simulations/startup.py
@@ -2,7 +2,7 @@
 # Authored by Robert Applin, 2020
 import sys
 
-from n_body_simulations.xml.interface_resources_rc import qInitResources, qCleanupResources
+from n_body_simulations.xml_r.interface_resources_rc import qInitResources, qCleanupResources
 from n_body_simulations.qt.model import NBodySimulationsModel
 from n_body_simulations.qt.presenter import NBodySimulationsPresenter
 from n_body_simulations.qt.view import NBodySimulationsView
